requirements.txt
tweepy
google-cloud-functions
google-cloud-bigquery
google-cloud-aiplatform
python-dotenv